Cleaned up classes, refactored singletons to static signatures, added tap-to focus.
